Move reduction specific fieldCollection methods to the Adaptor
